+++ requirements.txt
@@ -1,0 +1,5 @@
+MPU6050_light.h
+
+PubSubClient.h
+Arduinojson.h
+WifiClientSecure.h
